requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2020.12.5
Django==3.2
django-crispy-forms==1.11.2
django-decouple==2.0
djangorestframework==3.12.4
gunicorn==20.1.0
pycodestyle==2.7.0
python-decouple==3.4
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wincertstore==0.2